refactor: 공통 컴포넌트 스타일링 수정
