+++ src/components/layout/layout.tsx
@@ -19,4 +19,5 @@ const Container = styled.div`
   max-width: ${SIZE.width};
   margin: auto;
   box-sizing: border-box;
+  overflow: hidden;
 `;
