+++ src/components/layout/main.tsx
@@ -1,15 +1,18 @@
 import { COLORS } from "@/constants/css";
 import styled from "@emotion/styled";
-import AddressForm from "./address-form";
+import { ReactNode } from "react";
 
-const MAIN_TEXT = "만날 사람 주소를 추가해주세요";
+interface MainProps {
+  children: ReactNode;
+  text: string;
+}
 
-const Main = () => {
+const Main = ({ children, text }: MainProps) => {
   return (
     <MainContainer>
       <BorderContainer />
-      <TextP>{MAIN_TEXT}</TextP>
-      <AddressForm />
+      <TextP>{text}</TextP>
+      {children}
     </MainContainer>
   );
 };
@@ -20,6 +23,7 @@ const MainContainer = styled.main`
   width: 100%;
   max-height: 625px;
   height: 76vh;
+  position: relative;
   background-color: ${COLORS.backgroundPrimary};
   display: flex;
   flex-direction: column;
@@ -31,15 +35,17 @@ const MainContainer = styled.main`
 const TextP = styled.p`
   text-align: center;
   font-size: 1.2rem;
-  margin: 4rem 0 3.3rem 0;
+  margin-bottom: 3.3rem;
   opacity: 0.7;
   color: ${COLORS.semiBlack};
 `;
 
 const BorderContainer = styled.div`
+  position: absolute;
+  left: 0;
+  top: -15px;
   height: 25px;
   width: 100%;
   background-color: ${COLORS.backgroundPrimary};
-  margin-top: -15px;
   border-radius: 20px 20px 0 0;
 `;
